refactor(api): consolidate mentors/bookings/portfolios/config endpoints into catch-all routes
Collapses per-action serverless functions into one [...path].ts handler per
resource (bookings, mentors, portfolios) plus a shared admin-config
[resource].ts for topics/booking-rules/tools, staying under Vercel's function
count limit as the API surface grows.
